+++ FinancePro.UI/Views/ReportingView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace FinancePro.UI.Views;
+
+public partial class ReportingView : UserControl
+{
+    public ReportingView() => InitializeComponent();
+}